Adding DAC support in Managed SNI
