+++ src/System.Data.SqlClient/src/System/Data/SqlClient/SNI/SSRP.cs
@@ -1,0 +1,149 @@
+﻿using System.Diagnostics;
+using System.Net;
+using System.Net.Sockets;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace System.Data.SqlClient.SNI
+{
+    internal class SSRP
+    {
+        private const char SemicolonSeparator = ';';
+        private const int SqlServerBrowserPort = 1434;
+
+        /// <summary>
+        /// Finds instance port number for given instance name.
+        /// </summary>
+        /// <param name="browserHostname">SQL Sever Browser hostname</param>
+        /// <param name="instanceName">instance name to find port number</param>
+        /// <returns>port number for given instance name</returns>
+        internal static int GetPortByInstanceName(string browserHostName, string instanceName)
+        {
+            Debug.Assert(!string.IsNullOrWhiteSpace(browserHostName));
+            Debug.Assert(!string.IsNullOrWhiteSpace(instanceName));
+
+            byte[] instanceInfoRequest = CreateInstanceInfoRequest(instanceName);
+            byte[] responsePacket = SendUDPRequest(browserHostName, SqlServerBrowserPort, instanceInfoRequest);
+
+            const byte SvrResp = 0x05;
+            if (responsePacket == null || responsePacket.Length <= 3 || responsePacket[0] != SvrResp ||
+                BitConverter.ToUInt16(responsePacket, 1) != responsePacket.Length - 3)
+            {
+                throw new SocketException();
+            }
+
+            string serverMessage = Encoding.ASCII.GetString(responsePacket, 3, responsePacket.Length - 3);
+
+            string[] elements = serverMessage.Split(SemicolonSeparator);
+            int tcpIndex = Array.IndexOf(elements, "tcp");
+            if (tcpIndex < 0 || tcpIndex == elements.Length - 1)
+            {
+                throw new SocketException();
+            }
+
+            return ushort.Parse(elements[tcpIndex + 1]);
+        }
+
+        /// <summary>
+        /// Creates instance port lookup request (CLNT_UCAST_INST) for given instance name.
+        /// </summary>
+        /// <param name="instanceName">instance name to lookup port</param>
+        /// <returns>Byte array of instance port lookup request (CLNT_UCAST_INST)</returns>
+        private static byte[] CreateInstanceInfoRequest(string instanceName)
+        {
+            Debug.Assert(!string.IsNullOrWhiteSpace(instanceName));
+
+            const byte ClntUcastInst = 0x04;
+            int byteCount = Encoding.ASCII.GetByteCount(instanceName);
+
+            byte[] requestPacket = new byte[byteCount + 1];
+            requestPacket[0] = ClntUcastInst;
+            Encoding.ASCII.GetBytes(instanceName, 0, instanceName.Length, requestPacket, 1);
+
+            return requestPacket;
+        }
+
+        /// <summary>
+        /// Finds DAC port for given instance name.
+        /// </summary>
+        /// <param name="browserHostName">SQL Sever Browser hostname</param>
+        /// <param name="instanceName">instance name to lookup DAC port</param>
+        /// <returns>DAC port for given instance name</returns>
+        internal static int GetDacPortByInstanceName(string browserHostName, string instanceName)
+        {
+            Debug.Assert(!string.IsNullOrWhiteSpace(browserHostName));
+            Debug.Assert(!string.IsNullOrWhiteSpace(instanceName));
+
+            byte[] dacPortInfoRequest = CreateDacPortInfoRequest(instanceName);
+            byte[] responsePacket = SendUDPRequest(browserHostName, SqlServerBrowserPort, dacPortInfoRequest);
+
+            Console.WriteLine("responsePacket: " + BitConverter.ToString(responsePacket));
+
+            const byte SvrResp = 0x05;
+            const byte ProtocolVersion = 0x01;
+            const byte RespSize = 0x06;
+            if (responsePacket == null || responsePacket.Length <= 4 || responsePacket[0] != SvrResp ||
+                BitConverter.ToUInt16(responsePacket, 1) != RespSize || responsePacket[3] != ProtocolVersion)
+            {
+                throw new SocketException();
+            }
+
+            int dacPort = BitConverter.ToUInt16(responsePacket, 4);
+            return dacPort;
+        }
+
+        /// <summary>
+        /// Creates DAC port lookup request (CLNT_UCAST_DAC) for given instance name.
+        /// </summary>
+        /// <param name="instanceName">instance name to lookup DAC port</param>
+        /// <returns>Byte array of DAC port lookup request (CLNT_UCAST_DAC)</returns>
+        private static byte[] CreateDacPortInfoRequest(string instanceName)
+        {
+            Debug.Assert(!string.IsNullOrWhiteSpace(instanceName));
+
+            const byte ClntUcastDac = 0x0F;
+            const byte ProtocolVersion = 0x01;
+            int byteCount = Encoding.ASCII.GetByteCount(instanceName);
+
+            byte[] requestPacket = new byte[byteCount + 2];
+            requestPacket[0] = ClntUcastDac;
+            requestPacket[1] = ProtocolVersion;
+            Encoding.ASCII.GetBytes(instanceName, 0, instanceName.Length, requestPacket, 2);
+
+            return requestPacket;
+        }
+
+        /// <summary>
+        /// Sends request to server, and receives response from server by UDP.
+        /// </summary>
+        /// <param name="browserHostname">UDP server hostname</param>
+        /// <param name="port">UDP server port</param>
+        /// <param name="requestPacket">request packet</param>
+        /// <returns>response packet from UDP server</returns>
+        private static byte[] SendUDPRequest(string browserHostname, int port, byte[] requestPacket)
+        {
+            Debug.Assert(!string.IsNullOrWhiteSpace(browserHostname));
+            Debug.Assert(port >= 0 || port <= 65535);
+            Debug.Assert(requestPacket != null && requestPacket.Length > 0);
+
+            const int sendTimeOut = 1000;
+            const int receiveTimeOut = 1000;
+
+            IPAddress address = null;
+            IPAddress.TryParse(browserHostname, out address);
+
+            byte[] responsePacket = null;
+            using (UdpClient client = new UdpClient(address == null ? AddressFamily.InterNetwork : address.AddressFamily))
+            {
+                Task<int> sendTask = client.SendAsync(requestPacket, requestPacket.Length, browserHostname, port);
+                Task<UdpReceiveResult> receiveTask = null;
+                if (sendTask.Wait(sendTimeOut) && (receiveTask = client.ReceiveAsync()).Wait(receiveTimeOut))
+                {
+                    responsePacket = receiveTask.Result.Buffer;
+                }
+            }
+
+            return responsePacket;
+        }
+    }
+}
